# cmd/api/main.go
package main

import (
	"fmt"
	"log"
	"net/http"

	"ecopoints-go-api/internal/config"
	"ecopoints-go-api/internal/dto"
	"ecopoints-go-api/internal/handler"
	"ecopoints-go-api/internal/middleware"
	"ecopoints-go-api/internal/repository"
	"ecopoints-go-api/internal/service"

	_ "ecopoints-go-api/docs" // Swagger docs generated by swag

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

// @title EcoPoints Golang REST API
// @version 1.0
// @description REST API for EcoPoints application built with Go Gin and connected to existing Laravel MySQL database.
// @termsOfService http://swagger.io/terms/

// @contact.name EcoPoints Dev Team
// @contact.email support@ecopoints.test

// @license.name MIT
// @license.url https://opensource.org/licenses/MIT

// @host localhost:8080
// @BasePath /
// @schemes http https

// @securityDefinitions.apikey BearerAuth
// @in header
// @name Authorization
// @description Type "Bearer " followed by your JWT token. Example: "Bearer eyJhbGciOi..."

func main() {
	// 1. Load configuration
	cfg := config.LoadConfig()

	// 2. Set Gin mode
	if cfg.AppEnv == "production" {
		gin.SetMode(gin.ReleaseMode)
	} else {
		gin.SetMode(gin.DebugMode)
	}

	// 3. Connect to existing Laravel database
	db, err := config.InitDB(cfg)
	if err != nil {
		log.Fatalf("Failed to initialize database: %v", err)
	}

	// 4. Initialize Repositories
	userRepo := repository.NewUserRepository(db)
	wasteTypeRepo := repository.NewWasteTypeRepository(db)
	dropPointRepo := repository.NewDropPointRepository(db)
	rewardRepo := repository.NewRewardRepository(db)

	// 5. Initialize Services
	jwtService := service.NewJWTService(cfg.JWTSecret, cfg.JWTExpirationHours)
	authService := service.NewAuthService(userRepo, jwtService)
	wasteTypeService := service.NewWasteTypeService(wasteTypeRepo)
	dropPointService := service.NewDropPointService(dropPointRepo)
	rewardService := service.NewRewardService(rewardRepo)

	// 6. Initialize Handlers
	authHandler := handler.NewAuthHandler(authService)
	wasteTypeHandler := handler.NewWasteTypeHandler(wasteTypeService)
	dropPointHandler := handler.NewDropPointHandler(dropPointService)
	rewardHandler := handler.NewRewardHandler(rewardService)

	// 7. Initialize Gin Router
	router := gin.Default()

	// CORS Middleware
	router.Use(func(c *gin.Context) {
		c.Writer.Header().Set("Access-Control-Allow-Origin", "*")
		c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, PATCH, DELETE")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}

		c.Next()
	})

	// Health check endpoint
	router.GET("/health", func(c *gin.Context) {
		c.JSON(http.StatusOK, dto.SuccessResponse("EcoPoints Go API is healthy and connected to database", gin.H{
			"app_name": cfg.AppName,
			"database": cfg.DBDatabase,
		}))
	})

	// Swagger documentation route
	router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	// API v1 Routes
	v1 := router.Group("/api/v1")
	{
		// 1. Auth Routes
		authGroup := v1.Group("/auth")
		{
			authGroup.POST("/register", authHandler.Register)
			authGroup.POST("/login", authHandler.Login)
			authGroup.GET("/me", middleware.AuthMiddleware(jwtService), authHandler.Me)
		}

		// 2. Waste Types Routes (CRUD)
		wasteTypesGroup := v1.Group("/waste-types")
		{
			wasteTypesGroup.GET("", wasteTypeHandler.GetAll)
			wasteTypesGroup.GET("/:id", wasteTypeHandler.GetByID)

			// Admin/Petugas only write routes
			adminWaste := wasteTypesGroup.Group("")
			adminWaste.Use(middleware.AuthMiddleware(jwtService), middleware.RoleMiddleware("admin", "petugas"))
			{
				adminWaste.POST("", wasteTypeHandler.Create)
				adminWaste.PUT("/:id", wasteTypeHandler.Update)
				adminWaste.DELETE("/:id", wasteTypeHandler.Delete)
			}
		}

		// 3. Drop Points Routes (CRUD)
		dropPointsGroup := v1.Group("/drop-points")
		{
			dropPointsGroup.GET("", dropPointHandler.GetAll)
			dropPointsGroup.GET("/:id", dropPointHandler.GetByID)

			// Admin/Petugas only write routes
			adminDropPoints := dropPointsGroup.Group("")
			adminDropPoints.Use(middleware.AuthMiddleware(jwtService), middleware.RoleMiddleware("admin", "petugas"))
			{
				adminDropPoints.POST("", dropPointHandler.Create)
				adminDropPoints.PUT("/:id", dropPointHandler.Update)
				adminDropPoints.DELETE("/:id", dropPointHandler.Delete)
			}
		}

		// 4. Rewards Routes (CRUD)
		rewardsGroup := v1.Group("/rewards")
		{
			rewardsGroup.GET("", rewardHandler.GetAll)
			rewardsGroup.GET("/:id", rewardHandler.GetByID)

			// Admin only write routes
			adminRewards := rewardsGroup.Group("")
			adminRewards.Use(middleware.AuthMiddleware(jwtService), middleware.RoleMiddleware("admin"))
			{
				adminRewards.POST("", rewardHandler.Create)
				adminRewards.PUT("/:id", rewardHandler.Update)
				adminRewards.DELETE("/:id", rewardHandler.Delete)
			}
		}
	}

	// 8. Start server
	addr := fmt.Sprintf(":%s", cfg.AppPort)
	log.Printf("Server starting on port %s (Swagger docs: http://localhost:%s/swagger/index.html)", cfg.AppPort, cfg.AppPort)
	if err := router.Run(addr); err != nil {
		log.Fatalf("Server failed to start: %v", err)
	}
}